Move queries to separate files
