+++ asistente_ladm_col/lib/queries/annex_17_report/annex17_building_data_query.py
@@ -1,0 +1,18 @@
+def get_annex17_building_data_query (schema):
+    query = """SELECT array_to_json(array_agg(features)) AS features
+                    FROM (
+                    	SELECT f AS features
+                    	FROM (
+                    		SELECT 'Feature' AS type
+                    			,ST_AsGeoJSON(poligono_creado)::json AS geometry
+                    			,row_to_json((
+                    					SELECT l
+                    					FROM (
+                    						SELECT t_id AS t_id
+                    						) AS l
+                    					)) AS properties
+                            FROM {schema}.construccion AS c
+                    		) AS f
+                        ) AS ff;""".format(schema=schema)
+
+    return query
