+++ asistente_ladm_col/lib/queries/igac/basic_query.py
