+++ asistente_ladm_col/lib/queries/igac/economic_query.py
