+++ asistente_ladm_col/lib/queries/igac/legal_query.py
