+++ asistente_ladm_col/lib/queries/igac/physical_query.py
