+++ asistente_ladm_col/lib/queries/igac/property_record_card_query.py
